+++ pages/api/v1/auth/actions/addPostReaction.ts
@@ -62,16 +62,18 @@ async function handler(req: NextApiRequest, res: NextApiResponse<MessageType>) {
 
 	const subsquidProposalType  = getSubsquidLikeProposalType(postType);
 
-	if (!isNaN(trackNumber)){
-		// delete referendum v2 redis cache
-		if(postType == ProposalType.REFERENDUM_V2){
-			const trackListingKey = `${network}_${subsquidProposalType}_trackId_${trackNumber}_*`;
-			await deleteKeys(trackListingKey);
-		}
+	if (process.env.IS_CACHING_ALLOWED == '1'){
+		if (!isNaN(trackNumber)){
+			// delete referendum v2 redis cache
+			if(postType == ProposalType.REFERENDUM_V2){
+				const trackListingKey = `${network}_${subsquidProposalType}_trackId_${trackNumber}_*`;
+				await deleteKeys(trackListingKey);
+			}
 
-	}else if(postType == ProposalType.DISCUSSIONS){
-		const discussionListingKey = `${network}_${ProposalType.DISCUSSIONS}_page_*`;
-		await deleteKeys(discussionListingKey);
+		}else if(postType == ProposalType.DISCUSSIONS){
+			const discussionListingKey = `${network}_${ProposalType.DISCUSSIONS}_page_*`;
+			await deleteKeys(discussionListingKey);
+		}
 	}
 
 	await reactionsCollRef.doc(reactionDoc.id).set(reactionData, { merge: true }).then(() => {
